defines

+++ arduino/asserv/CC/serial_defines.h
@@ -52,8 +52,8 @@ enum ordre {
 	A_ROTR, 	//@float
 	A_KILLG, 
 	A_CLEANG,
-	A_PIDA, 	//@int @int @int
-	A_PIDD, 	//@int @int @int
+	A_PIDA, 	//@float @float @float
+	A_PIDD, 	//@float @float @float
 	A_GET_CODER, 	//#long #long
 	A_PWM_TEST, 	//@int @int @int
 	A_ACCMAX,	//@float
